auranode-api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
supabase==2.4.2
python-jose[cryptography]==3.3.0
groq==0.8.0
onnxruntime==1.18.0
Pillow==10.3.0
easyocr==1.7.1
apscheduler==3.10.4
slowapi==0.1.9
httpx==0.27.0
python-multipart==0.0.9
ruff==0.4.4
pytest==8.2.0
pytest-asyncio==0.23.6
requests==2.31.0
